run zola check on PRs; don't build on PR
this helps us ensure contributions are those where the site builds and
that the links are valid.

before the workflow for deploy was running and failing for
contributions. this will be more useful. plus we only want to deploy
when merged into `source`

in the future previews of PRs would be nice but this is better than
before